+++ test_requirements.txt
@@ -1,4 +1,4 @@
-tornado<5.0.0
+tornado>=5
 coverage[toml]
 twisted
 trio
