+++ urwid/tests/test_event_loops.py
@@ -114,6 +114,8 @@ else:
             from tornado.ioloop import IOLoop
             self.evl = urwid.TornadoEventLoop(IOLoop())
 
+        _expected_idle_handle = None
+
 
 try:
     import twisted
