+++ urwid/util.py
@@ -431,7 +431,7 @@ def _tagmarkup_recurse(tm, attr):
     return [tm], [(attr, len(tm))]
 
 
-def is_mouse_event(ev: str) -> bool:
+def is_mouse_event(ev: tuple[str, int, int, int] | typing.Any) -> bool:
     return isinstance(ev, tuple) and len(ev) == 4 and "mouse" in ev[0]
